fix: throttle event used during mode changed
ModeChanged is triggered **a lot** on some operation. This causes the
autocmd to be executed way too many times, slowing down some operations
like delete surrounding. One solution is to throttle it like I did in
this commit. An other one, that might be better, would be to reduce the
calls to `enable/disable` by either not having an autocmd on mode
changes, or to select some mode transition that are worth it.